connect header and footer and write 4 function to get product in homepage and detail
