+++ system/library/DController.php
@@ -2,8 +2,8 @@
 class DController {
     protected $load;
 
-public function __construct() {
-    $this->load = new Load();
-}
+    public function __construct() {
+        $this->load = new Load();
+    }
 
 }
